remove Endpoint directories
