+++ config/dependencies.global.php
@@ -37,10 +37,10 @@ use ZfrOAuth2\Server\Grant\AuthorizationGrant;
 use ZfrOAuth2\Server\Grant\ClientCredentialsGrant;
 use ZfrOAuth2\Server\Grant\PasswordGrant;
 use ZfrOAuth2\Server\Grant\RefreshTokenGrant;
-use ZfrOAuth2\Server\Middleware\Endpoint\AuthorizationRequestMiddleware;
-use ZfrOAuth2\Server\Middleware\Endpoint\RevocationRequestMiddleware;
-use ZfrOAuth2\Server\Middleware\Endpoint\TokenRequestMiddleware;
+use ZfrOAuth2\Server\Middleware\AuthorizationRequestMiddleware;
 use ZfrOAuth2\Server\Middleware\ResourceServerMiddleware;
+use ZfrOAuth2\Server\Middleware\RevocationRequestMiddleware;
+use ZfrOAuth2\Server\Middleware\TokenRequestMiddleware;
 use ZfrOAuth2\Server\Options\ServerOptions;
 use ZfrOAuth2\Server\ResourceServer;
 use ZfrOAuth2\Server\Service\AccessTokenService;
